refactor(User): move package user.domain to user.domain.user
