+++ src/main/java/com/sequence/anonymous/security/CustomOAuth2UserInfoFactory.java
@@ -4,7 +4,7 @@ import com.sequence.anonymous.security.userinfo.CustomOAuth2UserInfo;
 import com.sequence.anonymous.security.userinfo.GithubOAuth2UserInfo;
 import com.sequence.anonymous.security.userinfo.GoogleOAuth2UserInfo;
 import com.sequence.anonymous.security.userinfo.KakaoOAuth2UserInfo;
-import com.sequence.anonymous.user.domain.OAuth2Provider;
+import com.sequence.anonymous.user.domain.user.OAuth2Provider;
 import java.util.Map;
 
 public class CustomOAuth2UserInfoFactory {
