+++ src/main/java/com/sequence/anonymous/security/service/CustomOidcUserService.java
@@ -3,8 +3,8 @@ package com.sequence.anonymous.security.service;
 import com.sequence.anonymous.security.CustomOAuth2UserInfoFactory;
 import com.sequence.anonymous.security.CustomOidcUser;
 import com.sequence.anonymous.security.userinfo.CustomOAuth2UserInfo;
-import com.sequence.anonymous.user.domain.OAuth2Provider;
-import com.sequence.anonymous.user.domain.User;
+import com.sequence.anonymous.user.domain.user.OAuth2Provider;
+import com.sequence.anonymous.user.domain.user.User;
 import com.sequence.anonymous.user.domain.repository.UserRepository;
 import java.util.Map;
 import lombok.RequiredArgsConstructor;
