--- src/main/java/com/sequence/anonymous/user/domain/Gender.java
@@ -1,6 +1,0 @@
-package com.sequence.anonymous.user.domain;
-
-public enum Gender {
-  FEMALE,
-  MALE
-}
