+++ src/main/java/com/sequence/anonymous/user/domain/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.sequence.anonymous.user.domain.repository;
 
-import com.sequence.anonymous.user.domain.User;
+import com.sequence.anonymous.user.domain.user.User;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
