+++ src/main/java/com/sequence/anonymous/user/domain/user/Gender.java
@@ -1,0 +1,6 @@
+package com.sequence.anonymous.user.domain.user;
+
+public enum Gender {
+  FEMALE,
+  MALE
+}
